C: packet bool storage is uint8_t (#759)
Packet C stored bool as int (four bytes) while table C stored uint8_t.
A table of a packet type with two adjacent bools then failed its own
offsetof assert (firing at 4, layout says 1). Storage is always
uint8_t, so adding a table does not move packet headers.
serialize_read_bool still takes int *; the reader uses a temp and
assigns the field only after the read succeeds. Packet ABI of
bool-containing structs changes; wire bytes do not.

Closes #759. A child typed. Johnny Grok owns.

+++ testdata/golden/c/Wire.h
@@ -168,12 +168,12 @@ typedef struct ProbeBits {
 
 /* type ProbeSample */
 typedef struct ProbeSample {
-    int active;
+    uint8_t active;
     float orientation;
     int32_t raw_delta;
     int64_t big_delta;
     Weapon weapon;
-    int has_target;
+    uint8_t has_target;
     uint16_t target_id;
     uint32_t idle_ticks;
     uint16_t samples[8];
@@ -368,7 +368,7 @@ typedef struct TestData {
     uint32_t d;
     uint32_t e;
     uint32_t f;
-    int g;
+    uint8_t g;
     int32_t items[16];
     int32_t items_count;
     float float_value;
